# [core] Partially restore render performance