@alexshalamov 2019-07-29
@johnkassebaum Thanks for report and PR! We are looking into it at the moment. Our benchmarks do not show any regressions, therefore, performance drop could be related to the style being used. It would be helpful if you could provide more information. 

The best option would be to provide the style, if that is not possible, would be good to know how many sources and layers style has and how they interact while map is panned / zoomed (e.g., what are the expressions used for layer visibility, source zoom extents, etc.)