@pozdnyakov 2019-07-29
@johnkassebaum Thanks for your analysis and the PR! We've just made a patch, which moves shrinking of `filteredLayersForSource` out of the rendering call chain (https://github.com/mapbox/mapbox-gl-native/pull/15240). It might worth testing on top of it.

As for `layerIsVisible` and `zoomFitsLayer` computation, both are just numeric comparison, which makes me unsure how they could cause so significant FPS drop. It would be great to see the style that reproduces the issue. Also, this style could be part of our benchmarks and that would help to avoid regressions in future.