@pozdnyakov 2019-07-30
> We wish we could share our stylesheet but it represents a considerable investment. 

It is totally understandable, however having a benchmark with a mock style sheet would prevent us from experiencing similar issues in future. Would it be an option to add a test to https://github.com/mapbox/mapbox-gl-native/blob/master/benchmark/api/render.benchmark.cpp that illustrates performance benefit of the proposed changes? Think the key point is just having multiple sources with large amount of layers.

As I can see from the proposed changes, the mock style sheet should not depend on particular type of either layers or sources, therefore layers could be easily added using [runtime style API](https://github.com/mapbox/mapbox-gl-native/blob/master/include/mbgl/style/style.hpp#L69) in a cycle (as many as required).