@johnkassebaum 2019-08-06
No promise on a timeline to provide you with tests, but I can assure you that your refactored code is causing a significant performance hit on iOS and, as reported elsewhere, on Android. 

In the meantime, besides removing the memory re-allocations within the render pass loop, perhaps you could investigate removing repeated layer lookup work that is more often than not unnecessary due to the layer not being associated with the source of the outer loop iteration.