@johnkassebaum 2019-08-06
I also repeat the request for means of conducting and the results gathered on your own internal tests that showed no performance hit.  The iOS API has no current way of accurately measuring frame rate.