@pozdnyakov 2019-08-13
@johnkassebaum I made the following benchmark to try the proposed changes

```
static void API_renderStill_many_invisible_layers(::benchmark::State& state) {
    using namespace mbgl::style;
    RenderBenchmark bench;
    HeadlessFrontend frontend { size, pixelRatio };
    Map map { frontend, MapObserver::nullObserver(),
              MapOptions().withMapMode(MapMode::Static).withSize(size).withPixelRatio(pixelRatio),
              ResourceOptions().withCachePath(cachePath).withAccessToken("foobar") };
    prepare(map);
    auto& style = map.getStyle();
    auto source = std::make_unique<GeoJSONSource>("GeoJSONSource");
    style.addSource(std::move(source));
    for (int i = 0; i < 10000; ++i) {
        auto layer = std::make_unique<SymbolLayer>(std::to_string(i), "GeoJSONSource");
        layer->setVisibility(VisibilityType::None);
        style.addLayer(std::move(layer));
    }

    while (state.KeepRunning()) {
        frontend.render(map);
    }
}
```

Unfortunately, it did not show any significant performance changes. Could you please check whether this test shows performance benefit in your environment, or fix the test, so that it better simulates your style data.

> I also repeat the request for means of conducting and the results gathered on your own internal benchmarks (" Our benchmarks do not show any regressions") that showed no performance hit.

I can assure you that we take performance and UI responsiveness very seriously and use various metrics to check it on all the supported platforms. However, the engine performance depends a lot on the tile data and styles being used and this might explain the difference of what we experience.