@pozdnyakov 2020-02-06
Performance for the styles that contain lots of sources had been improved with https://github.com/mapbox/mapbox-gl-native/pull/15756. Please feel free to rebase&reopen this PR if you've measurements showing that it still makes a performance benefit.